+++ source/extensions/lightspeed.trex.utils.widget/lightspeed/trex/utils/widget/__init__.py
@@ -21,16 +21,18 @@ from .asset_validation import (
     open_replacement_asset_file_picker,
 )
 from .categories_dialog.categories_dialog import RemixCategoriesDialog
-from .invalid_deps_dialog import show_invalid_deps_rebuild_dialog
-from .message_dialog import TrexMessageDialog
+from .invalid_deps_dialog import confirm_invalid_deps_rebuild, show_invalid_deps_rebuild_dialog
+from .message_dialog import MessageDialogResult, TrexMessageDialog
 from .workspace import WorkspaceWidget, WorkspaceWindowBase
 
 __all__ = [
+    "MessageDialogResult",
     "RemixCategoriesDialog",
     "TrexMessageDialog",
     "WorkspaceWidget",
     "WorkspaceWindowBase",
     "accept_asset_if_valid_for_replacement",
+    "confirm_invalid_deps_rebuild",
     "open_asset_file_picker",
     "open_replacement_asset_file_picker",
     "show_invalid_deps_rebuild_dialog",